Add (rural) pages to Website
The first commit to the Odin Recipe Book.
To represent a light version of a recipe-website and further
establish knowledge of HTML, the Odin Project assigns to build
a webpage, consisting of "index.html" and three (3) different recipes.

As for the content, it was up to the student (me), to fill in the blanks.
I used a recipe of the hyperlinked Website "https://www.allrecipes.com/",
the latter two are copied from "https://biancazapatka.com/de/".

The shown images refer to the original picture and are not downloaded.
The index has a link to each Recipe, wherefrom a button an the bottom
sends you back to the Homepage.