Create default Top-Navigation-Bar
As flexboxes were introduced, creating a,
now good-looking navigation-bar is a lot easier.
The Site-Title was transformed into a link to
redirect to the index.html, as for the links to the
individual recipes.
There's also an image on the main-page to fill the gap.

Individual recipes have to be styled now; also it needs
to be figured out, how to shrink down items to
their container, without managing to calculate the
remaining space of it (e.g. image on top-nav-bar)